On-screen graphic finder
Finds the location of an element in a GUI interface whatever the screen resolution and wherever this element is located.
The script takes a full screenshot, which helps locate the item you want.
It is sometimes necessary to add the parameters x and y of the command to move the automatic click of the mouse which allows to interact with the element or therefore to select it.
The script can be modified to use interfaces other than GNOME and other screenshots software.
Example command: `./auto-click-screenreader.sh allow_button 50 30`
Sources : https://blog.sleeplessbeastie.eu/2013/01/21/how-to-automate-mouse-and-keyboard/